requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
diskcache==5.6.3
fastapi==0.112.4
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gradio==4.43.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
llama_cpp_python==0.2.90
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
ruff==0.7.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
